+++ CHANGELOG.md
@@ -6,10 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ## Unrealesed
 ### Added
 - License file : the licensing hasn't been changed but has been made more explicited with the adequate section in the readme and the licenses files
+- New mechanism to adapt a user model frm closure through the `UserModelFromClosure` and `UserModelFromClosureAndJacobian` structs.
 ### Changed
 - Improve documentation and rewrite of the README.md
 - *Breaking change*: Define API thanks to rexport, access to `newton_rootfinder::solver_advanced` has been deleted, use directly `newton_rootfinder` from now on.
-- *Breaking change*: Use `UserModelFromFunc` and `UserModelFromFuncAndJacobian` instead of `UserModelWithFunc` and `UserModelWithFuncJac`
+- *Breaking change*: Use `UserModelFromFunction` and `UserModelFromFunctionAndJacobian` instead of `UserModelWithFunction` and `UserModelWithFunctionJac`
 - Moved from nalgebra 0.21.1 to 0.26.2
 
 ### Removed